Minor cleanup; move templates to a separate file; prepare for splitting template parsing and execution
